+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.18.2] - 2025-12-05
+
 ### Added
 
 - `show-all-tasks.sh` - display a list of all running ECS tasks (even non-caccl-deploy ones)